Added more API endpoints, moved types to a dedicated folder
