+++ src/types/Events.ts
@@ -1,0 +1,13 @@
+export type Events = {
+  boss: {
+    name: string,
+    timestamp: number
+  },
+  helltide: {
+    timestamp: number,
+    refresh: number
+  },
+  legion: {
+    timestamp: number
+  }
+};
