+++ src/types/Modes.ts
@@ -1,0 +1,10 @@
+/**
+ * The mode type.
+ * @type {Mode}
+ */
+export type Modes = 'allmodes' | 'softcore' | 'hardcore' | 'dead' | 'pvp';
+
+export type ModesChoices = {
+  name: string;
+  value: Modes;
+};
