+++ src/types/index.ts
@@ -1,0 +1,9 @@
+export * from './Player';
+export * from './Modes';
+export * from './Events';
+export * from './Classes';
+
+export type Status = {
+	character_service: boolean,
+	event_service: boolean
+}
